requirements.txt
requests>=2.0.0
gTTS>=2.0.0
python-dotenv>=0.19.0
discord.py>=2.0.0
